link_api/app/auth/auth.py
from datetime import datetime, timezone, timedelta
from jose import jwt
from pwdlib import PasswordHash

from app.config.project_config import settings


password_hash = PasswordHash.recommended()

def get_password_hash(password: str) -> str:
    return password_hash.hash(password)

def verify_password(plain_password: str, hashed_password: str) -> bool:
    return password_hash.verify(plain_password, hashed_password)

def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.now(timezone.utc) + timedelta(minutes=30)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(
        claims=to_encode,
        key=settings.SECRET_KEY,
        algorithm=settings.SERVICE_ALGORITHM
    )
    return encoded_jwt

