link_api/pyproject.toml
[project]
name = "link-api"
version = "0.1.0"
description = ""
authors = [
    {name = "andreychvankov",email = "whitebear134@gmail.com"}
]
requires-python = "~3.13"
dependencies = [
    "fastapi (>=0.128.7,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "aiohttp (>=3.13.3,<4.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg (>=3.3.2,<4.0.0)",
    "sqladmin (>=0.23.0,<0.24.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "pyright[nodejs] (>=1.1.408,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "autoflake (>=2.3.1,<3.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "pyright (>=1.1.408,<2.0.0)"
]

[tool.black]
line-length = 88
target-version = ['py313']

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.autoflake]
check_diff = true
imports = ["sqlalchemy", "app"]

[tool.pyright]
include = ["app"]